+++ Ranger/UI/PlayersTableViewController/PlayersTableView/Cells/Player/PlayerNameCellView.swift
@@ -17,11 +17,23 @@ class PlayerNameCellView: PlayerCellView
     // UI.
     @IBOutlet weak var flagImageView: NSImageView!
     
+    // Data.
+    var countries: Countries = Countries.load()
     
+    
+    override func layout()
+    {
+        super.layout()
+        
+        // Some hardcoded UI.
+        flagImageView.layer?.cornerRadius = 2.0
+        flagImageView.alphaValue = 0.5
+    }
+        
     override func prepareForReuse()
     {
         // UI.
-        flagImageView.image = nil
+        flagImageView.image = NSImage(named: "Unknown")
     }
     
     override func setup(with player: Model.Player, in tableColumn: NSTableColumn?)
@@ -42,7 +54,20 @@ class PlayerNameCellView: PlayerCellView
             let country = player.sharkScope.summary?.Response.PlayerResponse.PlayerView.Player.country,
             let countryName = player.sharkScope.summary?.Response.PlayerResponse.PlayerView.Player.countryName
         {
-            flagImageView.toolTip = countryName
+            // Estimate country date.
+            let now = Date()
+            let timezone = countries.country(for: country.uppercased())?.averageTimezone
+            let countryDateFormatter = DateFormatter()
+                countryDateFormatter.dateFormat = "MMM d, H:mm"
+                countryDateFormatter.timeZone = timezone
+            let dateString = countryDateFormatter.string(from: now)
+            
+            print(now)
+            print(timezone)
+            print(countryDateFormatter)
+            print(dateString)
+            
+            flagImageView.toolTip = "\(countryName) (\(dateString))"
             flagImageView.image = NSImage(named: country.uppercased())
         }
     }