+++ CHANGELOG.md
@@ -4,7 +4,6 @@
 * Doing
 
     + Next up
-        + Add Local time to right-click context menu
         + UI for Opted-Out SharkScope players
         + Aggregate `Tournaments` sessions / calculate session stats
         
@@ -22,6 +21,12 @@
         + Normalized / Absolute switch for Finishes histogram
         + `TableWindowInfo`
             + Regex title recognition (with tests)
+
+* 2.0.0
+
+    + Added Local time to country flag tooltip
+        + Created `Countries` (tool for estimate local time for country)
+        + Covered `Countries` with tests
 
 * 2.0.0